# Conflicts:
#	target/classes/com/example/oj/controller/QuestionController.class
#	target/classes/com/example/oj/domain/entity/User.class
#	target/classes/com/example/oj/service/IQuestionService.class
#	target/classes/com/example/oj/service/impl/QuestionServiceImpl.class

+++ src/main/java/com/example/oj/domain/entity/QuestionTag.java
@@ -25,14 +25,14 @@ public class QuestionTag implements Serializable {
 
     private static final long serialVersionUID = 1L;
 
-   @TableId(value = "id", type = IdType.AUTO)
+    @TableId(value = "id", type = IdType.AUTO)
     private Integer id;
 
     @TableField("tag_id")
-    private Integer tagId;
+    private Long tagId;
 
     @TableField("question_id")
-    private Integer questionId;
+    private Long questionId;
 
 
 }
